Compiler 2.0 26th pass on rewrite - Init Processing started - return statements processed - if statements processed - expression statements processed - added control path evaluation (this shit is freakin siiiicckkkk) - Label address remapping added - for loops processed - aliases processed - foreach statements processed - while statements processed

+++ lib/grammar/backend/data/DataEntity.cpp
@@ -6,4 +6,4 @@
 #include "../../frontend/parser/Ast.h"
 #include "../Compiler.h"
 
-long invalidAddr = -109912;
+int invalidAddr = -109912;
